// src/payload-types.ts
export interface PricingBlock {
  header?: string | null;
  title: string;
  description?: string | null;
  plans: {
    product: number | Product;
    id?: string | null;
  }[];
  id?: string | null;
  blockName?: string | null;
  blockType: 'pricing';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "products".
 */
export interface Product {
  id: number;
  stripeProductID: string;
  name?: string | null;
  description?: string | null;
  features?:
    | {
        name?: string | null;
        id?: string | null;
      }[]
    | null;
  metadata?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  image?: string | null;
  prices?:
    | {
        priceId?: string | null;
        currency?: string | null;
        type?: string | null;
        recurringInterval?: string | null;
        amount?: number | null;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}